\name{birdring-package}
\alias{birdring-package}
\docType{package}
\title{Methods to analyse bird ring reencounter data}
\description{
This package is a collection of functions or methods used to analyse ring reencounter data.
Its purpose is to read EURING ring reencounter data into R, and to help mapping and analysing reencounter data.
}
\details{
\tabular{ll}{
Package: \tab birdring\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-17\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Fraenzi Korner-Nievergelt and Robert Robinson

Maintainer: Fraenzi Korner-Nievergelt <fraenzi.korner@vogelwarte.ch>
}
\references{
Speek, G., Clark, J.A., Rhode, Z., Wassenaar, R.D. & van Noordwijk, A.J. (2001) The EURING exchange-code 2000. Dutch Ringing Scheme, Heteren.
}
\keyword{ package }
\keyword{ ring recovery }
\keyword{ ring reencounter data }
\keyword{ euring }
