\name{lancol}
\alias{lancol}
\docType{data}
\title{
Example data set of EURING2000 format
}
\description{
Ringing and recovery data of the Red-backed Shrike in the EURING2000 format
}
\usage{data(lancol)}
\format{
  A data frame with 2483 observations of the 40 variables of the EURING2000.
}
\source{
http://www.euring.org
}
\references{
Speek, G., Clark, J.A., Rhode, Z., Wassenaar, R.D. & van Noordwijk, A.J. (2001) The EURING exchange-code 2000. Dutch Ringing Scheme, Heteren.
}
\examples{
data(lancol)
}
\keyword{datasets}
