\name{species}
\alias{species}
\docType{data}
\title{
EURING-code for species
}
\description{
EURING-code for species as given in Speek et al. (2001)
}
\usage{data(species)}
\format{
  A data frame with 3311 observations on the following 3 variables.
  \describe{
    \item{\code{Code}}{a numeric vector with the EURING species code}
    \item{\code{Name}}{a factor with the species names}
    \item{\code{Date.Updated}}{a factor}
}
}
\source{
http://www.euring.org
}
\references{
Speek, G., Clark, J.A., Rhode, Z., Wassenaar, R.D. & van Noordwijk, A.J. (2001) The EURING exchange-code 2000. Dutch Ringing Scheme, Heteren.
}
\examples{
data(species)
}
\keyword{datasets}
