\name{birdring-package}
\alias{birdring-package}
\docType{package}
\title{Methods to Analyse Bird Ring Reencounter data}
\description{
This package is a collection of functions or methods used to analyse ring reencounter data.
Its purpose is to read EURING ring reencounter data into R, and to help mapping and analysing reencounter data.
For a tutorial, go to http://www.tandfonline.com/doi/full/10.1080/03078698.2014.933053
}
\details{
\tabular{ll}{
Package: \tab birdring\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-09-24\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Fraenzi Korner-Nievergelt and Robert Robinson

Maintainer: Fraenzi Korner-Nievergelt <fraenzi.korner@vogelwarte.ch>
}
\references{
Speek, G., Clark, J.A., Rhode, Z., Wassenaar, R.D. & van Noordwijk, A.J. (2001) The EURING exchange-code 2000. Dutch Ringing Scheme, Heteren.
}
\keyword{ package }
\keyword{ ring recovery }
\keyword{ ring reencounter data }
\keyword{ euring }
